Update index.js file
Add the function to check if url starts with http.
if the url start with http, continue the function with the url
if not, add the http to the url.
And get request from the url,
if error is exist, consolo.log the error and res.send Down!.
if error is not exist and response status code less than and under to 445, response send up!
If not, response send Down!.